finalyearproject/requirements.txt
Django==5.0.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0



#pip install -r requirements.txt